Attempts to normalise spacing.
This commit attempts to normalise some of the spacing used by various files
throughout the codebase (this includes things like stripping out
unnecessary trailing whitespace, and changing anomalous spacing between
function declarations and statements, and the braces that come after,
the spacing used after commas and variable declarations, etc to the styles
most commonly already used by the files where the anomalous spacing
exists). However, it isn't thorough, and further potential work in this
area could be performed.

This commit attempts to slightly normalise the code style already used by
most of the codebase. However, it isn't thorough, and further potential
work in this area could be performed.

This commit doesn't add/remove any functionality/features, and shouldn't
change the behaviour of anything in the codebase.