include as statement instead of function.
Because include is a statement (as opposed to a function), it's better to
use it in a statement-like way, instead of a function-like way (although
either way will still work all the same). This commit normalises include
usage to be statement-like rather than function-like.

Also spotted/fixed another small typo and did some more braces and spaces
work.